Some minor mistakes fixed and tidy up
A few poorly written lines have been fixed but animations still do not
work. I've found a large problem with the keyframe times however. The
mTime values for the animNode in the scenes animations channels are are
exactly the same so there is no way for the code to know what keyframes
to use for the transforms. This could entirely be the reason for it not
working but the only thing online refering the this problem are on
assimp.sourceforge which is currently offline. Animations will need to
be on hold until this issue can be fixed