AI Pathfinding Added
The A* algorithm has been added. This includes the algorithm itself
(aStarPathFinder), a graph representing all nodes/paths in a scene
(aStarGraph) and a node class. An AI System class has been set up to
handle to handle the A* algorithm and an AI_Components will use this AI
system.