Speed Tweaked.
AI objects work well. They have positions stored that represent
tergets/incentives. Right now these are just vec3(0,0,-50) ,
vec3(50,0,0) etc but should be defined as things like the player
position or places to stand and defend. When a target is selected the A*
algorithm find theshortest path and the object follows this. Right now
the incentives are loop through but an algorithm is needed to allow the
object to decide what it's next incentive should be