Bullet Shooting
The raycast class and new bulletParticle are now tied together. When R
is pressed the raycast class calls on bullerParticle 'emit' function
which requires a velocity and position parameter. Currently supports one
bulet which is re-emitted when R is pressed. To Do: support multiple
bullets and allow changing of particle colour.