HUD Health Bar
A working HUD health bar. PlayerHUD class may now be deprecated. All HUD
item information is in HUDitem which has it's own renderer it uses.
HUDitem is a component attached to the player.

TO DO:
-Needs to discard texture transparancy in the frag shader.
-Design Good HUD textures
-Link textures to real time game values eg health changes which then
change the texture.
-Intended to use orthographic projection to render the 2D HUD since it's
the proper way but ran into issues and it now uses perspective
projection in a cheating kind of manner which would be good to fix at
some point