Changing HUD and Design Change
The HUD is now linked to It's user. A new class, HUD_Logic is abtract
and represents what a HUD's logic should do which is deciding on
texture, position and scaling changes. I've defined a HUD_Health logic
class which controls the health bar. When you make a HUD component it
require to be given a HUD_Logic object. HUD_Component now just stores
core data (position, scaling and texture) and rendering itself while
it's HUD_Logic object hides away all the decisions and tests that
changes rely on.