State Machine Foundation
In order to introduce menus next, we need some sort of state machine. A
new "Iron Rifts" class is going to store and control all the possible
states of the game. Right now there's one state. There's a lot to do
before actually getting to the menu since the input management needs
completely over hauled. "Iron Rifts" will now be the only place inputs
are polled and it will then poll inputs for state change triggers and
then pass input into state for the states to process. this should fix
the input lags and allow for different state eg game mode and menus to
be added flexibly. Hopefully. Right now the original "Game" class which
held and updated all the object that made up our deathmatch game has now
become the state "InfiniteDeathmatchState" and is held in our new Iron
Rifts class which should essentially act as a state machine.