Progress with State Machine and In Game Menu
State managment function added to IronRifts and our gameplay state is
able to toggle a state change via a key press. Started creating an
InGameMenu state which can then be used to test and debug any issues
with the state management.

I also took all the code from main that involved loading the current
game mode we have and tucked it into a seperate file called FristLevel,
creatively, for the sake of cleaning main.