State changes and InGame Menu
The renderer that was being used for the HUD has been abstracted to
accept anything implementing the abstract class "UIRenderable" which
defines the minimal information needed to draw the image. A new class
called MenuUI defines a menu screen and is attached to a game object
that a new InGameMenu state can update. So we have the game pausing and
switching to and from the play state and menu. Because the game is
running at a high rate a timers been added to the states so we can delay
state change tests because it was closing and opening the menu
instantaneously on key presses. Currently key "P" toggles the in game
menu on and off.

This menu is currently just an image and to start with I'll make a
subMenu which gives access to the leaderboard to view it.