Start, pause and playstate working
There is now a simple start screen upon starting game and a pause menu.
There was a big bug that was causing only one of the objects using
UIRenderer to acrtually render. Not sure what exactly it was but
UIRenderer wasnt loading the shader i thought it was and correcting this
fixed the bug. Next thing will be adding some functionality to the
menus, at a minimum this would include being able to access the
leaderboard through the pause menu and possibly a level selection in the
start menu. Could do with better textures for all this as well.